16 bits is not sufficient to hold size of binary encoded to transfer to propeller bootloader

+++ src/propellerloader.h
@@ -49,7 +49,7 @@ namespace PZST {
         uint8_t iterateLFSR();
         void sendByte(uint8_t);
         void sendLong(uint32_t);
-        void sendBytes(char*, uint16_t);
+        void sendBytes(char*, int);
         bool getReply(char*c, int timeout);
         void advance(int);
